# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rooms": {
      "get": {
        "operationId": "Rooms_list",
        "parameters": [
          {
            "name": "buildingCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "CS",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "ค้นจากรหัสหรือชื่อห้อง",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "roomType",
            "required": false,
            "in": "query",
            "description": "ประเภทห้อง: LAB = ห้องปฏิบัติการคอมพิวเตอร์ · LECTURE = ห้องบรรยาย · NETWORK_LAB = ห้องปฏิบัติการเครือข่าย · MEETING = ห้องประชุม · OFFICE = ห้องพัก/สำนักงาน · OTHER = อื่นๆ",
            "schema": {
              "$ref": "#/components/schemas/RoomType"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true = เฉพาะห้องที่เปิดใช้งาน",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RoomDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ห้องทั้งหมด (กรองตามอาคาร) พร้อมสรุปสถานะเครื่องในห้อง",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "operationId": "Rooms_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/RoomDetailDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_UNAVAILABLE — ติดต่อ Core Hub ไม่ได้ชั่วคราว (มี header Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "เพิ่มห้อง (ระบบสุ่มรหัส QR ของห้องให้)",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{id}": {
      "get": {
        "operationId": "Rooms_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/RoomDetailDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ห้อง + เครื่องทั้งหมดในห้องพร้อมสถานะ (แจ้งแล้ว / กำลังซ่อม / รออะไหล่)",
        "tags": [
          "rooms"
        ]
      },
      "patch": {
        "operationId": "Rooms_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/RoomDetailDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_UNAVAILABLE — ติดต่อ Core Hub ไม่ได้ชั่วคราว (มี header Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "แก้ไขห้อง / เปิด-ปิดการใช้งาน",
        "tags": [
          "rooms"
        ]
      },
      "delete": {
        "operationId": "Rooms_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ลบห้องพร้อมเครื่องในห้อง — ใบแจ้งซ่อมเดิมยังอยู่เป็นประวัติ (ไม่ผูกห้อง/เครื่องแล้ว)",
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/api/v1/rooms/{id}/equipment": {
      "post": {
        "operationId": "Rooms_addEquipment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEquipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/RoomDetailDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "เพิ่มเครื่องในห้อง (ใส่ count เพื่อเพิ่มหลายเครื่อง เช่น PC-01 ถึง PC-30)",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/equipment": {
      "get": {
        "operationId": "Rooms_inventory",
        "parameters": [
          {
            "name": "buildingCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "CS",
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "ประเภทอุปกรณ์ (หมวดหมู่งานซ่อม)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "ค้นจากป้าย ชื่อ/รุ่น หรือเลขครุภัณฑ์",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true = เฉพาะเครื่องที่เปิดใช้งาน",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EquipmentInventoryDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ทะเบียนครุภัณฑ์ — เครื่องทุกห้อง (กรองตามอาคาร/ห้อง/ประเภท/คำค้น) ไม่แบ่งหน้า สูงสุด 2000 แถว",
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/api/v1/qr-codes/{code}": {
      "get": {
        "operationId": "Rooms_resolveQr",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/QrTargetDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "สแกน QR — รหัสนี้เป็นของห้องหรือเครื่องไหน",
        "tags": [
          "rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "csmju-maintenance-request API",
    "description": "ระบบแจ้งซ่อม — ระบบย่อยของ CSMJU2030 · ทุก endpoint ใต้ /api/v1 ต้องมี Core Hub access token (Authorization: Bearer หรือคุกกี้ csmju_maintenance_request_access_token) · response ห่อด้วย envelope มาตรฐาน",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256 จาก Core Hub)",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "RoomType": {
        "type": "string",
        "enum": [
          "LAB",
          "LECTURE",
          "NETWORK_LAB",
          "MEETING",
          "OFFICE",
          "OTHER"
        ],
        "description": "ประเภทห้อง: LAB = ห้องปฏิบัติการคอมพิวเตอร์ · LECTURE = ห้องบรรยาย · NETWORK_LAB = ห้องปฏิบัติการเครือข่าย · MEETING = ห้องประชุม · OFFICE = ห้องพัก/สำนักงาน · OTHER = อื่นๆ"
      },
...
      "RoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "buildingCode": {
            "type": "string",
            "example": "CS"
          },
          "building": {
            "$ref": "#/components/schemas/BuildingRefDto"
          },
          "code": {
            "type": "string",
            "example": "CS-201"
          },
          "name": {
            "type": "string",
            "example": "ห้องปฏิบัติการคอมพิวเตอร์ 1"
          },
          "floor": {
            "type": "number",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "roomType": {
            "description": "ประเภทห้อง: LAB = ห้องปฏิบัติการคอมพิวเตอร์ · LECTURE = ห้องบรรยาย · NETWORK_LAB = ห้องปฏิบัติการเครือข่าย · MEETING = ห้องประชุม · OFFICE = ห้องพัก/สำนักงาน · OTHER = อื่นๆ",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomType"
              }
            ]
          },
          "capacity": {
            "type": "number",
            "nullable": true,
            "description": "จำนวนที่นั่ง"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true,
            "example": "/api/v1/rooms/0d6f…/photo?v=1759480000000",
            "description": "URL รูปห้อง (?v= เปลี่ยนเมื่อแก้ไข) · null = ยังไม่มีรูป"
          },
          "qrCode": {
            "type": "string",
            "description": "รหัสใน QR ของห้อง (/q/<qrCode>)"
          },
          "isActive": {
            "type": "boolean"
          },
          "equipmentStates": {
            "description": "สถานะเครื่องทั้งหมดในห้อง",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomStateSummaryDto"
              }
            ]
          },
          "openRequestCount": {
            "type": "number",
            "description": "ใบแจ้งซ่อมของห้องนี้ที่ยังไม่ปิด (รวมที่ไม่ได้ผูกเครื่อง)"
          }
        },
        "required": [
          "id",
          "buildingCode",
          "building",
          "code",
          "name",
          "floor",
          "description",
          "roomType",
          "capacity",
          "photoUrl",
          "qrCode",
          "isActive",
          "equipmentStates",
          "openRequestCount"
        ]
      },
...
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "buildingCode": {
            "type": "string",
            "example": "CS",
            "description": "code ของอาคารใน Core Hub"
          },
          "code": {
            "type": "string",
            "example": "CS-201",
            "description": "A–Z ตัวเลข และ - เท่านั้น"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 150,
            "example": "ห้องปฏิบัติการคอมพิวเตอร์ 1"
          },
          "floor": {
            "type": "number",
            "nullable": true,
            "minimum": -5,
            "maximum": 99
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "roomType": {
            "description": "ประเภทห้อง: LAB = ห้องปฏิบัติการคอมพิวเตอร์ · LECTURE = ห้องบรรยาย · NETWORK_LAB = ห้องปฏิบัติการเครือข่าย · MEETING = ห้องประชุม · OFFICE = ห้องพัก/สำนักงาน · OTHER = อื่นๆ",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomType"
              }
            ]
          },
          "capacity": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 1000,
            "description": "จำนวนที่นั่ง"
          }
        },
        "required": [
          "buildingCode",
          "code",
          "name"
        ]
      },
...
    }
  }
}